+++ hemi/internal/web_relay_uwsgi.go
@@ -3,7 +3,7 @@
 // All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE.md file.
 
-// UWSGI proxy implementation.
+// UWSGI relay implementation.
 
 // UWSGI is mainly for Python applications. See: https://uwsgi-docs.readthedocs.io/en/latest/Protocol.html
 // UWSGI 1.9.13 seems to have unsized content support: https://uwsgi-docs.readthedocs.io/en/latest/Chunked.html
@@ -17,23 +17,23 @@ import (
 )
 
 func init() {
-	RegisterHandlet("uwsgiProxy", func(name string, stage *Stage, app *App) Handlet {
-		h := new(uwsgiProxy)
+	RegisterHandlet("uwsgiRelay", func(name string, stage *Stage, app *App) Handlet {
+		h := new(uwsgiRelay)
 		h.onCreate(name, stage, app)
 		return h
 	})
 }
 
-// uwsgiProxy handlet passes web requests to backend uWSGI servers and cache responses.
-type uwsgiProxy struct {
+// uwsgiRelay handlet passes web requests to backend uWSGI servers and cache responses.
+type uwsgiRelay struct {
 	// Mixins
 	Handlet_
 	contentSaver_ // so responses can save their large contents in local file system.
 	// Assocs
 	stage   *Stage      // current stage
-	app     *App        // the app to which the proxy belongs
+	app     *App        // the app to which the relay belongs
 	backend wireBackend // the *TCPSBackend or *TUDSBackend to pass to
-	cacher  Cacher      // the cacher which is used by this proxy
+	cacher  Cacher      // the cacher which is used by this relay
 	// States
 	bufferClientContent bool          // client content is buffered anyway?
 	bufferServerContent bool          // server content is buffered anyway?
@@ -42,16 +42,16 @@ type uwsgiProxy struct {
 	maxContentSize      int64         // max response content size allowed
 }
 
-func (h *uwsgiProxy) onCreate(name string, stage *Stage, app *App) {
+func (h *uwsgiRelay) onCreate(name string, stage *Stage, app *App) {
 	h.MakeComp(name)
 	h.stage = stage
 	h.app = app
 }
-func (h *uwsgiProxy) OnShutdown() {
+func (h *uwsgiRelay) OnShutdown() {
 	h.app.SubDone()
 }
 
-func (h *uwsgiProxy) OnConfigure() {
+func (h *uwsgiRelay) OnConfigure() {
 	h.contentSaver_.onConfigure(h, TmpsDir()+"/web/uwsgi/"+h.name)
 	// toBackend
 	if v, ok := h.Find("toBackend"); ok {
@@ -61,13 +61,13 @@ func (h *uwsgiProxy) OnConfigure() {
 			} else if wireBackend, ok := backend.(wireBackend); ok {
 				h.backend = wireBackend
 			} else {
-				UseExitf("incorrect backend '%s' for uwsgiProxy, must be TCPSBackend or TUDSBackend\n", name)
+				UseExitf("incorrect backend '%s' for uwsgiRelay, must be TCPSBackend or TUDSBackend\n", name)
 			}
 		} else {
 			UseExitln("invalid toBackend")
 		}
 	} else {
-		UseExitln("toBackend is required for uwsgiProxy")
+		UseExitln("toBackend is required for uwsgiRelay")
 	}
 
 	// withCacher
@@ -112,14 +112,14 @@ func (h *uwsgiProxy) OnConfigure() {
 		return errors.New(".maxContentSize has an invalid value")
 	}, _1T)
 }
-func (h *uwsgiProxy) OnPrepare() {
+func (h *uwsgiRelay) OnPrepare() {
 	h.contentSaver_.onPrepare(h, 0755)
 }
 
-func (h *uwsgiProxy) IsProxy() bool { return true }
-func (h *uwsgiProxy) IsCache() bool { return h.cacher != nil }
+func (h *uwsgiRelay) IsProxy() bool { return true }
+func (h *uwsgiRelay) IsCache() bool { return h.cacher != nil }
 
-func (h *uwsgiProxy) Handle(req Request, resp Response) (next bool) { // reverse only
+func (h *uwsgiRelay) Handle(req Request, resp Response) (next bool) { // reverse only
 	// TODO: implementation
 	resp.Send("uwsgi")
 	return
@@ -128,7 +128,7 @@ func (h *uwsgiProxy) Handle(req Request, resp Response) (next bool) { // reverse
 // poolUWSGIExchan
 var poolUWSGIExchan sync.Pool
 
-func getUWSGIExchan(proxy *uwsgiProxy, conn wireConn) *uwsgiExchan {
+func getUWSGIExchan(relay *uwsgiRelay, conn wireConn) *uwsgiExchan {
 	var exchan *uwsgiExchan
 	if x := poolUWSGIExchan.Get(); x == nil {
 		exchan = new(uwsgiExchan)
@@ -139,7 +139,7 @@ func getUWSGIExchan(proxy *uwsgiProxy, conn wireConn) *uwsgiExchan {
 	} else {
 		exchan = x.(*uwsgiExchan)
 	}
-	exchan.onUse(proxy, conn)
+	exchan.onUse(relay, conn)
 	return exchan
 }
 func putUWSGIExchan(exchan *uwsgiExchan) {
@@ -154,7 +154,7 @@ type uwsgiExchan struct {
 	response uwsgiResponse // the uwsgi response
 }
 
-func (x *uwsgiExchan) onUse(proxy *uwsgiProxy, conn wireConn) {
+func (x *uwsgiExchan) onUse(relay *uwsgiRelay, conn wireConn) {
 }
 func (x *uwsgiExchan) onEnd() {
 }